refactor: ClassPage 컴포넌트의 비지니스 로직을 useClassPage 훅으로 분리 #180
